Refactor: separate Phase 1 data ingestion from Phase 2 logic
